+++ src/Tabs/TabList.jsx
@@ -2,7 +2,7 @@ import React, { Component, PropTypes, createElement } from 'react'
 import specialAssign from '../utils/special-assign'
 
 const checkedProps = {
-  component: PropTypes.string,
+  component: PropTypes.any,
   children:  PropTypes.oneOfType([PropTypes.func, PropTypes.node]).isRequired
 }