+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Change prop name `tag` -> `component`
 
 Allow all `Manager` style components to just pass `children` through
 
+Make sure the `Overlay` component only returns focus if another focusable item was not clicked
+
 ### 0.9.0
 Move `Overlay` into `Overlays` namespace